old-tests/submission/W3C/domhtml/HTMLMetaElement04.html

<!DOCTYPE html>
<html><head>
<meta scheme="NIST" charset="utf-8">
<title>NIST DOM HTML Test - META</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-meta-scheme" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<br>
</p>



<script>
var domtest = async_test("The scheme IDL attribute of the meta element must reflect the content attribute of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vscheme;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("meta");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vscheme = testNode.scheme;

      assert_equals(vscheme,"NIST");
       
    domtest.done();
}));

</script>
<p>The scheme IDL attribute of the meta element must reflect the content attribute of the same name.</p>
<div id="log"></div>
</body></html>
